Move iterators into utils
