+++ libraries/wavemap/include/wavemap/data_structure/chunked_ndtree/chunked_ndtree.h
@@ -8,7 +8,7 @@
 #include "wavemap/common.h"
 #include "wavemap/data_structure/chunked_ndtree/ndtree_node_chunk.h"
 #include "wavemap/indexing/ndtree_index.h"
-#include "wavemap/iterator/subtree_iterator.h"
+#include "wavemap/utils/iterate/subtree_iterator.h"
 
 namespace wavemap {
 template <typename NodeDataT, int dim, int chunk_height>
