+++ libraries/wavemap/include/wavemap/integrator/ray_tracing/ray_tracing_integrator.h
@@ -6,7 +6,7 @@
 #include "wavemap/data_structure/volumetric/volumetric_data_structure_base.h"
 #include "wavemap/integrator/integrator_base.h"
 #include "wavemap/integrator/measurement_model/constant_ray.h"
-#include "wavemap/iterator/ray_iterator.h"
+#include "wavemap/utils/iterate/ray_iterator.h"
 
 namespace wavemap {
 /**
