+++ libraries/wavemap/include/wavemap/utils/iterate/grid_iterator.h
@@ -1,5 +1,5 @@
-#ifndef WAVEMAP_ITERATOR_GRID_ITERATOR_H_
-#define WAVEMAP_ITERATOR_GRID_ITERATOR_H_
+#ifndef WAVEMAP_UTILS_ITERATE_GRID_ITERATOR_H_
+#define WAVEMAP_UTILS_ITERATE_GRID_ITERATOR_H_
 
 #include <utility>
 
@@ -66,4 +66,4 @@ class Grid {
 };
 }  // namespace wavemap
 
-#endif  // WAVEMAP_ITERATOR_GRID_ITERATOR_H_
+#endif  // WAVEMAP_UTILS_ITERATE_GRID_ITERATOR_H_
