+++ libraries/wavemap/include/wavemap/utils/iterate/pointcloud_iterator.h
@@ -1,5 +1,5 @@
-#ifndef WAVEMAP_ITERATOR_POINTCLOUD_ITERATOR_H_
-#define WAVEMAP_ITERATOR_POINTCLOUD_ITERATOR_H_
+#ifndef WAVEMAP_UTILS_ITERATE_POINTCLOUD_ITERATOR_H_
+#define WAVEMAP_UTILS_ITERATE_POINTCLOUD_ITERATOR_H_
 
 #include <utility>
 
@@ -52,4 +52,4 @@ class PointcloudIterator {
 };
 }  // namespace wavemap
 
-#endif  // WAVEMAP_ITERATOR_POINTCLOUD_ITERATOR_H_
+#endif  // WAVEMAP_UTILS_ITERATE_POINTCLOUD_ITERATOR_H_
