+++ libraries/wavemap/src/integrator/projective/fixed_resolution/fixed_resolution_integrator.cc
@@ -1,7 +1,7 @@
 #include "wavemap/integrator/projective/fixed_resolution/fixed_resolution_integrator.h"
 
 #include "wavemap/indexing/index_conversions.h"
-#include "wavemap/iterator/grid_iterator.h"
+#include "wavemap/utils/iterate/grid_iterator.h"
 
 namespace wavemap {
 void FixedResolutionIntegrator::importPointcloud(
