+++ ros/wavemap_ros/src/input_handler/depth_image_input_handler.cc
@@ -3,7 +3,7 @@
 #include <cv_bridge/cv_bridge.h>
 #include <opencv2/core/eigen.hpp>
 #include <tracy/Tracy.hpp>
-#include <wavemap/iterator/grid_iterator.h>
+#include <wavemap/utils/iterate/grid_iterator.h>
 #include <wavemap/utils/print/eigen.h>
 
 namespace wavemap {
